[Store] Change keep_count parameter from int to size_t for type safety
- Changed CleanupOldSnapshots parameter from int to size_t in MasterSnapshotRepository
- Changed CleanupOldSnapshot parameter from int to size_t in MasterSnapshotManager
- Removed unnecessary static_cast operations when comparing with size()
- Improves type safety as keep_count represents a count that cannot be negative

+++ mooncake-store/include/master_snapshot_manager.h
@@ -104,7 +104,7 @@ class MasterSnapshotManager {
         const std::vector<uint8_t>& data, const std::string& path,
         const std::string& local_filename, const std::string& snapshot_id);
 
-    void CleanupOldSnapshot(int keep_count, const std::string& snapshot_id);
+    void CleanupOldSnapshot(size_t keep_count, const std::string& snapshot_id);
 
     std::string FormatTimestamp(
         const std::chrono::system_clock::time_point& tp);
